move encoding test at its implementation
